git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery@356 57fcd75e-5312-0410-8df3-f5eb6fbb1595

+++ delete_album.php
@@ -19,12 +19,16 @@
  */
 ?>
 <?
+// Hack check
+if (!$user->canDeleteAlbum($album)) {
+	exit;
+}
+
 if ($confirm) {
 	$album->delete();
 	dismissAndReload();
 	return;
 }
-
 
 if ($album) {
 ?>
